Revert "Moved some files to Components package"
This reverts commit cc904200b87bb2441d52c1dcc9addbf83566e81f.
