+++ app/src/main/java/com/belia/speedotransfer/ui/common_ui/TopBar.kt
@@ -1,4 +1,4 @@
-package com.belia.speedotransfer.ui.auth.components
+package com.belia.speedotransfer.ui.common_ui
 
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -14,7 +14,6 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import com.belia.speedotransfer.R
-import com.belia.speedotransfer.ui.theme.titleMedium
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -28,7 +27,8 @@ fun TopBar(
         title = {
             Text(
                 text = title,
-                style = titleMedium,
+                style = MaterialTheme.typography.titleMedium,
+                fontSize = 20.sp,
             )
         },
         navigationIcon = {
